feat: remove `archive_file` resource deprecation notice
- from what i gathered from the git history, this resource was
introduced in `v2.3.0` in a deprecated state

- discussion with the community involving @bendbennett resulted in the
consideration of this deprecation being removed (see: #218)

- since `v2.x` there has been a strong focus on keeping this provider
limited to the `data` source, so some of this proposed change is taking
a best guess at what makes sense while trying not to depart from the
overall approach

- i acknowledge that this appears to be in some ways orthogonal to the
goal of the project, but there has been little movement in regards to
this widely-requested feature (understandably so, based on reasons
mentioned above).

- there are cases, primarily in a remote build server context, where the
design decisions/limitations of the `data` source makes the adoption and
usage of this provider much more challenging.

this change is proposed as a stop-gap until a larger focus is able to
put on this project by the primary contributing team